+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/ast/plugin/PluginTupleField.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017, 2019 Reece H. Dunn
+ * Copyright (C) 2017, 2019-2020 Reece H. Dunn
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,13 +17,13 @@ package uk.co.reecedunn.intellij.plugin.xpath.ast.plugin
 
 import com.intellij.psi.PsiElement
 import uk.co.reecedunn.intellij.plugin.xdm.types.XdmSequenceType
-import uk.co.reecedunn.intellij.plugin.xdm.types.XsNCNameValue
+import uk.co.reecedunn.intellij.plugin.xdm.types.XsStringValue
 
 /**
  * A Saxon 9.8 `TupleField` node in the XQuery AST.
  */
 interface PluginTupleField : PsiElement {
-    val fieldName: XsNCNameValue
+    val fieldName: XsStringValue
 
     val fieldType: XdmSequenceType?